Split court order creation logic into new service
This significantly simplifies the OrgService service, and makes it easier to identify the logic being implemented by CourtOrderCreator
